Cover new feature with tests
